+++ Irvin.Fludal/IProcedureTarget.cs
@@ -1,0 +1,6 @@
+﻿namespace Irvin.Fludal;
+
+public interface IProcedureTarget<TSelf>
+{
+    public TSelf WithOutputParameter<T>(string parameterName);
+}
